+++ Configuration/TCA/Overrides/tt_content.php
@@ -44,4 +44,23 @@ call_user_func(function($namespace, $extkey) {
   $GLOBALS['TCA']['tt_content']['types']['list']['subtypes_excludelist'][$contentElementSignature] = 'layout,select_key,pages,recursive';
   $GLOBALS['TCA']['tt_content']['types']['list']['subtypes_addlist'][$contentElementSignature] = 'pi_flexform';
   \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPiFlexFormValue($contentElementSignature, 'FILE:EXT:' . $extkey . '/Configuration/FlexForms/ContentElements/QuickNavigationContentElement.xml');
+
+  // --- Subpages Navigation Content Element ---
+
+  \TYPO3\CMS\Extbase\Utility\ExtensionUtility::registerPlugin(
+    $extensionSignature,
+
+    // A unique name of the content element in upper camel case
+    'SubpagesNavigation',
+
+    // Description of the content element shown in the backend dropdown field
+    'Subpages Navigation Content Element'
+  );
+
+  $contentElementName      = strtolower('SubpagesNavigation');
+  $contentElementSignature = $extensionName . '_' . $contentElementName;
+
+  $GLOBALS['TCA']['tt_content']['types']['list']['subtypes_excludelist'][$contentElementSignature] = 'layout,select_key,pages,recursive';
+  $GLOBALS['TCA']['tt_content']['types']['list']['subtypes_addlist'][$contentElementSignature] = 'pi_flexform';
+  \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPiFlexFormValue($contentElementSignature, 'FILE:EXT:' . $extkey . '/Configuration/FlexForms/ContentElements/SubpagesNavigationContentElement.xml');
 }, 't3v', 't3v_navigations');