+++ Configuration/TSconfig/ContentElements/SubpagesNavigationContentElement.ts
@@ -1,0 +1,27 @@
+mod {
+  wizards {
+    newContentElement {
+      wizardItems {
+        contentElements {
+          elements {
+            t3vnavigations_subpagesnavigation {
+              title = LLL:EXT:t3v_navigations/Resources/Private/Language/locallang_wizard.xlf:contentElements.subpagesNavigationContentElement.title
+
+              description = LLL:EXT:t3v_navigations/Resources/Private/Language/locallang_wizard.xlf:contentElements.subpagesNavigationContentElement.description
+
+              iconIdentifier = t3v_navigations-subpages_navigation_content_element
+
+              tt_content_defValues {
+                CType = list
+
+                list_type = t3vnavigations_subpagesnavigation
+              }
+            }
+          }
+
+          show := addToList(t3vnavigations_subpagesnavigation)
+        }
+      }
+    }
+  }
+}