+++ ext_localconf.php
@@ -46,6 +46,26 @@ call_user_func(function($namespace, $extkey) {
     ]
   );
 
+  // --- Subpages Navigation Content Element ---
+
+  \TYPO3\CMS\Extbase\Utility\ExtensionUtility::configurePlugin(
+    $extensionSignature,
+
+    // A unique name of the content element in upper camel case
+    'SubpagesNavigation',
+
+    // An array holding the controller-action-combinations that are accessible
+    [
+      // The first controller and its first action will be the default
+      'ContentElements\SubpagesNavigationContentElement' => 'index'
+    ],
+
+    // An array of non-cachable controller-action-combinations (they must already be enabled)
+    [
+      // 'ContentElements\SubpagesNavigationContentElement' => 'index'
+    ]
+  );
+
   // === TsConfig ===
 
   \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPageTSConfig('<INCLUDE_TYPOSCRIPT: source="FILE:EXT:' . $extkey . '/Configuration/TSconfig.ts">');